feat: implement ChatAgentPromptProvider with default context prompt logic
