+++ packages/startup/entry/sample-modules/ai-native/ai-native.contribution.ts
@@ -11,7 +11,6 @@ import {
 import { TextWithStyle } from '@opensumi/ide-ai-native/lib/browser/contrib/terminal/utils/ansi-parser';
 import {
   AINativeCoreContribution,
-  ChatAgentPromptProvider,
   ERunStrategy,
   IChatFeatureRegistry,
   IInlineChatFeatureRegistry,
@@ -26,6 +25,10 @@ import {
 } from '@opensumi/ide-ai-native/lib/browser/types';
 import { InlineChatController } from '@opensumi/ide-ai-native/lib/browser/widget/inline-chat/inline-chat-controller';
 import { SerializedContext } from '@opensumi/ide-ai-native/lib/common/llm-context';
+import {
+  ChatAgentPromptProvider,
+  DefaultChatAgentPromptProvider,
+} from '@opensumi/ide-ai-native/lib/common/prompts/context-prompt-provider';
 import { MergeConflictPromptManager } from '@opensumi/ide-ai-native/lib/common/prompts/merge-conflict-prompt';
 import { RenamePromptManager } from '@opensumi/ide-ai-native/lib/common/prompts/rename-prompt';
 import { TerminalDetectionPromptManager } from '@opensumi/ide-ai-native/lib/common/prompts/terminal-detection-prompt';
@@ -500,36 +503,9 @@ export class AINativeContribution implements AINativeCoreContribution {
   }
 
   registerChatAgentPromptProvider(): void {
-    this.injector.addProviders({
+    this.injector.overrideProviders({
       token: ChatAgentPromptProvider,
-      useValue: {
-        provideContextPrompt: (context: SerializedContext, userMessage: string) => `
-          <additional_data>
-          Below are some potentially helpful/relevant pieces of information for figuring out to respond
-          <recently_viewed_files>
-          ${context.recentlyViewFiles.map((file, idx) => `${idx} + 1: ${file}`)}
-          </recently_viewed_files>
-          <attached_files>
-          ${context.attachedFiles.map(
-            (file) =>
-              `
-          <file_contents>
-          \`\`\`${file.language} ${file.path}
-          ${file.content}
-          \`\`\`
-          </file_contents>
-          <linter_errors>
-          ${file.lineErrors.join('`n')}
-          </linter_errors>
-          `,
-          )}
-          
-          </attached_files>
-          </additional_data>
-          <user_query>
-          ${userMessage}
-          </user_query>`,
-      },
+      useValue: DefaultChatAgentPromptProvider,
     });
   }
 }
